updated project structure
- files are now in the right folders according to the angular style guide
