+++ src/app/movies/movie-details/movie-details.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { Movie } from '../movie';
+import { Movie } from '../shared/movie';
 import { ActivatedRoute } from '@angular/router';
-import { MovieService } from '../movie.service';
+import { MovieService } from '../shared/movie.service';
 
 @Component({
   selector: 'app-movie-details',
