+++ src/app/movies/movie-sidescroller/movie-sidescroller.component.spec.ts
